fixup! refactor(ng-dev): centralize the configuration/definition of labels that are controlled by tooling
