+++ ng-dev/pr/common/BUILD.bazel
@@ -3,7 +3,10 @@ load("//tools:defaults.bzl", "ts_library")
 ts_library(
     name = "common",
     srcs = glob(["**/*.ts"]),
-    visibility = ["//ng-dev:__subpackages__"],
+    visibility = [
+        "//github-actions/commit-message-based-labels:__subpackages__",
+        "//ng-dev:__subpackages__",
+    ],
     deps = [
         "//ng-dev/commit-message",
         "//ng-dev/pr/config",
