+++ ng-dev/pr/config/BUILD.bazel
@@ -4,7 +4,6 @@ ts_library(
     name = "config",
     srcs = ["index.ts"],
     visibility = [
-        "//github-actions/commit-message-based-labels/lib:__pkg__",
         "//ng-dev:__subpackages__",
     ],
     deps = [
